requirements.txt
python-telegram-bot
requests
urllib3
schedule
psycopg2